+++ CHANGELOG.md
@@ -5,6 +5,8 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 ## Unreleased
 
 - Updated to js-yaml v4 (breaking change of `safeLoad()` to just `load()`).
+- Include 'rdfs:Resource' as a vocab term property (should be rarely, if ever,
+  needed, but no harm in adding).
 
 ## 1.0.1 2021-08-20